ml-ops-2/interface/requirements.txt
streamlit
psycopg2-binary
pandas
matplotlib